[FIX] #84 - 성취뷰 API 수정
